# Optimize session auth tag generation