I profiled my application and crypto/hmac.New was taking 5.71% of the
total CPU time. Since the session key does not change, you can call
crypto/hmac.Reset() instead of making a new object. This reuses previous
work done for the key.

```
name                 old time/op  new time/op  delta
EncryptRTP-8         2.17µs ± 4%  1.90µs ± 5%  -12.79%  (p=0.008 n=5+5)
EncryptRTPInPlace-8  2.14µs ± 4%  1.84µs ± 4%  -13.79%  (p=0.008 n=5+5)
```