@kixelated 2019-02-16
btw, `DecryptRTCP` doesn't check the auth tag.