@Sean-Der 2019-02-16
@kixelated 

Blah thanks for catching that, would be awesome if you could fix + add tests for that also! Thanks so much